% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_zinarp.R
\name{explore_zinarp}
\alias{explore_zinarp}
\title{EXPLORATORY DATA ANALYSIS FOR ZINAR(p) PROCESSES}
\usage{
explore_zinarp(x)
}
\arguments{
\item{x}{A vector containing a discrete non-negative time series data set.}
}
\value{
Plot time series graph, relative frequency bar plot, autocorrelation function graph and partial autocorrelation function graph on a common plot.
}
\description{
This function generates a graph for exploring ZINAR(p) processes.
}
