% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slesions.R
\docType{data}
\name{slesions}
\alias{slesions}
\title{Skin lesions dataset}
\format{
An object of class \code{data.frame} with 84 rows and 1 columns.
}
\usage{
slesions
}
\description{
Monthly number of skin lesions-related submissions to animal health laboratories
from a region in New Zealand, obtained from 2003 to 2009.
}
\references{
Jazi, Mansour Aghababaei, Geoff Jones, and Chin‐Diew Lai. "First‐order integer valued AR processes with zero inflated Poisson innovations." Journal of Time Series Analysis 33.6 (2012): 954-963.
}
\keyword{datasets}
